+++ Resources/config/services/article.php
@@ -9,6 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
+use Sulu\Bundle\FormBundle\TitleProvider\ArticelTitleProvider;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 use function Symfony\Component\DependencyInjection\Loader\Configurator\service;
@@ -17,7 +18,7 @@ return static function(ContainerConfigurator $container) {
     $services = $container->services();
     $parameters = $container->parameters();
 
-    $services->set('sulu_form.title_provider.article', Sulu\Bundle\FormBundle\TitleProvider\StructureTitleProvider::class)
+    $services->set('sulu_form.title_provider.article', Sulu\Bundle\FormBundle\TitleProvider\ArticelTitleProvider::class)
         ->args([service('request_stack')])
         ->tag('sulu_form.title_provider', ['alias' => 'article']);
 };
