+++ Resources/config/services/title-providers.php
@@ -9,8 +9,10 @@
  * with this source code in the file LICENSE.
  */
 
+use Sulu\Bundle\FormBundle\TitleProvider\TitleProviderPoolInterface;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
+use Symfony\Component\HttpFoundation\RequestStack;
 use function Symfony\Component\DependencyInjection\Loader\Configurator\service;
 use function Symfony\Component\DependencyInjection\Loader\Configurator\tagged_iterator;
 
@@ -20,8 +22,13 @@ return static function(ContainerConfigurator $container) {
 
     $services->set('sulu_form.title_provider.pool', Sulu\Bundle\FormBundle\TitleProvider\TitleProviderPool::class)
         ->args([tagged_iterator('sulu_form.title_provider', indexAttribute: 'alias')]);
+    $services->alias(TitleProviderPoolInterface::class, 'sulu_form.title_provider.pool');
 
-    $services->set('sulu_form.title_provider.page', Sulu\Bundle\FormBundle\TitleProvider\StructureTitleProvider::class)
-        ->args([service('request_stack')])
-        ->tag('sulu_form.title_provider', ['alias' => 'page']);
+    $services->set('sulu_form.title_provider.page', Sulu\Bundle\FormBundle\TitleProvider\ArticleTitleProvider::class)
+        ->args([
+            service(RequestStack::class)
+        ])
+        ->tag('sulu_form.title_provider', ['alias' => 'structure']) # For BC reasons
+        ->tag('sulu_form.title_provider', ['alias' => 'article'])
+    ;
 };
